requirements.txt
PyQt6>=6.6.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
camelot-py[cv]>=0.11.0
anthropic>=0.40.0
openpyxl>=3.1.0
Pillow>=10.0.0
pyyaml>=6.0
opencv-python>=4.10
scikit-image>=0.24
numpy>=1.26
ultralytics>=8.3
